Use base class for BMX IX battery
